Add configuration system
The new configuration system, using TOML, loads into `app.config`
using a TOML config file. This can be specified on the command-line
with the `-c` option, or it can be at one of the following default
locations:

- Linux: $XDG_CONFIG_HOME/wuzz/config.toml (by default this is ~/.config/wuzz/config.toml)
- Other OSs: ~/.wuzz/config.toml

Currently the following keys can be set, along with their default
values:

```toml
[general]

timeout = "1m" # string parsed into time.Duration
formatJSON = true # toggles JSON formatting
preserveScrollPosition = true # currently unused
defaultURLScheme = "https" # when a URL is not provided, this is the scheme used by default
```

In addition, a [keys] section can also be set and will be successfully
parsed, but it is currently non-functional.