Support PATCH requests with JSON bodies
Chrome's 'Copy as cURL' option generates "-X PATCH -H 'Content-Type:
application/json' --data-binary '<json...>'" options to construct a
PATCH request that contains a JSON body.  This commit allows wuzz to
send such a request by making the following changes:

Modify App.ParseArgs to support the "--data-binary" option.  When this
option is specified, do not pass the body through url.QueryUnescape
and do not add automatically a "Content-Type:
application/x-www-form-urlencoded" header to the request.

Modify App.ParseArgs to remember if the request method has been
explicitly set via the "-X" or "--request" options and in such
situations to not automatically set the request method to POST when
request data is present.

Modify App.SubmitRequest to include the request body if the method is
PATCH.

Modify App.SubmitRequest to not replace newlines (\n) with
ampersands (&) in the request body unless a "Content-Type:
application/x-www-form-urlencoded" header exists.  This required
moving the creation of the headers to be before the creation of the
body and the actual request.